E/RecyclerView: No adapter attached; skipping layout

+++ app/src/main/java/com/example/moviedetails/ui/moviedetails/MovieDetailsFragment.kt
@@ -28,7 +28,7 @@ class MovieDetailsFragment : Fragment() {
 
     companion object {
         private const val MOVIE_ID_KEY = "MOVIE_ID_KEY"
-        fun newInstance(movieId: Long) = MovieDetailsFragment().apply {
+        fun newInstance(movieId: Int) = MovieDetailsFragment().apply {
             arguments = bundleOf(MOVIE_ID_KEY to movieId)
         }
     }
